+++ app/src/main/java/com/sergiolopez/runningpacecalculator/viewModel/DataViewModel.kt
@@ -1,6 +1,5 @@
 package com.sergiolopez.runningpacecalculator.viewModel
 
-import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.sergiolopez.runningpacecalculator.model.RunPaceModel
@@ -8,9 +7,9 @@ import com.sergiolopez.runningpacecalculator.model.RunPaceModel
 class DataViewModel : ViewModel() {
 
     val distanceSelected = MutableLiveData<Int>(0)
-    var hours = MutableLiveData<Float>(0f)
-    var minutes = MutableLiveData<Float>(0f)
-    var seconds = MutableLiveData<Float>(0f)
+    var hours = MutableLiveData<Double>(0.0)
+    var minutes = MutableLiveData<Double>(0.0)
+    var seconds = MutableLiveData<Double>(0.0)
     var resultPace = MutableLiveData<Double>(0.0)
     var splitTimesList = MutableLiveData<MutableList<Float>>()
 
@@ -18,23 +17,23 @@ class DataViewModel : ViewModel() {
         distanceSelected.value = distance
     }
 
-    fun setHours(hours: Float) {
+    fun setHours(hours: Double) {
         this.hours.value = hours
     }
 
-    fun setMinutes(minutes: Float) {
+    fun setMinutes(minutes: Double) {
         this.minutes.value = minutes
     }
 
-    fun setSeconds(seconds: Float) {
+    fun setSeconds(seconds: Double) {
         this.seconds.value = seconds
     }
 
     private fun createRunPaceObject(): RunPaceModel {
         val distance = distanceSelected.value ?: 0
-        val hoursValue = hours.value ?: 0f
-        val minutesValue = minutes.value ?: 0f
-        val secondsValue = seconds.value ?: 0f
+        val hoursValue = hours.value ?: 0.0
+        val minutesValue = minutes.value ?: 0.0
+        val secondsValue = seconds.value ?: 0.0
 
         return RunPaceModel(distance, hoursValue, minutesValue, secondsValue)
     }
